backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1
django-mptt>=0.14
Pillow>=10.0
